Major refactoring:
- Move distributed methods out of non-distributed modules
- Refactor dispatching
- Refactor structure of distributed modules
- Add classes for execution engine and memory format
