+++ awswrangler/_config.py
@@ -1,10 +1,8 @@
 """Configuration file for AWS SDK for pandas."""
 
-import importlib.util
 import inspect
 import logging
 import os
-from enum import Enum
 from typing import Any, Callable, Dict, List, NamedTuple, Optional, Tuple, Type, Union, cast
 
 import botocore.config
@@ -16,20 +14,6 @@ _logger: logging.Logger = logging.getLogger(__name__)
 
 
 _ConfigValueType = Union[str, bool, int, botocore.config.Config, None]
-
-
-class ExecutionEngine(Enum):
-    """Execution engine enum."""
-
-    RAY = "ray"
-    PYTHON = "python"
-
-
-class MemoryFormat(Enum):
-    """Memory format enum."""
-
-    MODIN = "modin"
-    PANDAS = "pandas"
 
 
 class _ConfigArg(NamedTuple):
@@ -69,12 +53,6 @@ _CONFIG_ARGS: Dict[str, _ConfigArg] = {
     "botocore_config": _ConfigArg(dtype=botocore.config.Config, nullable=True),
     "verify": _ConfigArg(dtype=str, nullable=True, loaded=True),
     # Distributed
-    "execution_engine": _ConfigArg(
-        dtype=str, nullable=False, loaded=True, default="ray" if importlib.util.find_spec("ray") else "python"
-    ),
-    "memory_format": _ConfigArg(
-        dtype=str, nullable=False, loaded=True, default="modin" if importlib.util.find_spec("modin") else "pandas"
-    ),
     "address": _ConfigArg(dtype=str, nullable=True),
     "redis_password": _ConfigArg(dtype=str, nullable=True),
     "ignore_reinit_error": _ConfigArg(dtype=bool, nullable=True),
@@ -439,24 +417,6 @@ class _Config:  # pylint: disable=too-many-instance-attributes,too-many-public-m
     @verify.setter
     def verify(self, value: Optional[str]) -> None:
         self._set_config_value(key="verify", value=value)
-
-    @property
-    def execution_engine(self) -> str:
-        """Property execution_engine."""
-        return cast(str, self["execution_engine"])
-
-    @execution_engine.setter
-    def execution_engine(self, value: str) -> None:
-        self._set_config_value(key="execution_engine", value=value)
-
-    @property
-    def memory_format(self) -> str:
-        """Property memory_format."""
-        return cast(str, self["memory_format"])
-
-    @memory_format.setter
-    def memory_format(self, value: str) -> None:
-        self._set_config_value(key="memory_format", value=value)
 
     @property
     def address(self) -> Optional[str]:
