+++ awswrangler/distributed/__init__.py
@@ -1,17 +1,1 @@
 """Distributed Module."""
-
-from awswrangler.distributed._distributed import (  # noqa
-    RayLogger,
-    initialize_ray,
-    modin_repartition,
-    ray_get,
-    ray_remote,
-)
-
-__all__ = [
-    "RayLogger",
-    "initialize_ray",
-    "modin_repartition",
-    "ray_get",
-    "ray_remote",
-]
