+++ awswrangler/distributed/ray/__init__.py
@@ -1,1 +1,10 @@
 """Ray Module."""
+
+from awswrangler.distributed.ray._core import RayLogger, initialize_ray, ray_get, ray_remote  # noqa
+
+__all__ = [
+    "RayLogger",
+    "initialize_ray",
+    "ray_get",
+    "ray_remote",
+]
