+++ awswrangler/distributed/ray/_core.pyi
@@ -14,7 +14,6 @@ class RayLogger:
 
 def ray_get(futures: List[Any]) -> List[Any]: ...
 def ray_remote(function: Callable[..., Any]) -> Callable[..., Any]: ...
-def modin_repartition(function: Callable[..., Any]) -> Callable[..., Any]: ...
 def initialize_ray(
     address: Optional[str] = None,
     redis_password: Optional[str] = None,
@@ -25,4 +24,3 @@ def initialize_ray(
     cpu_count: Optional[int] = None,
     gpu_count: Optional[int] = 0,
 ) -> None: ...
-def _get_ray_object_store_memory() -> Optional[int]: ...
