+++ awswrangler/distributed/ray/modin/__init__.py
@@ -1,0 +1,7 @@
+"""Ray Modin Module."""
+
+from awswrangler.distributed.ray.modin._core import modin_repartition
+
+__all__ = [
+    "modin_repartition",
+]
