+++ awswrangler/s3/_delete.py
@@ -9,7 +9,7 @@ import boto3
 
 from awswrangler import _utils, exceptions
 from awswrangler._threading import _get_executor
-from awswrangler.distributed import RayLogger, ray_get, ray_remote
+from awswrangler.distributed.ray import RayLogger, ray_get, ray_remote
 from awswrangler.s3._fs import get_botocore_valid_kwargs
 from awswrangler.s3._list import _path2list
