+++ awswrangler/s3/_write_text.py
@@ -11,26 +11,13 @@ import pandas as pd
 from pandas.io.common import infer_compression
 
 from awswrangler import _data_types, _utils, catalog, exceptions, lakeformation
-from awswrangler._config import ExecutionEngine, MemoryFormat, apply_configs, config
-from awswrangler.distributed import modin_repartition
+from awswrangler._config import apply_configs
+from awswrangler._distributed import engine
+from awswrangler.distributed.ray.modin import modin_repartition
 from awswrangler.s3._delete import delete_objects
 from awswrangler.s3._fs import open_s3_object
 from awswrangler.s3._write import _COMPRESSION_2_EXT, _apply_dtype, _sanitize, _validate_args
 from awswrangler.s3._write_dataset import _to_dataset
-
-if config.execution_engine == ExecutionEngine.RAY.value:
-    from ray.data import from_modin, from_pandas
-    from ray.data.datasource.file_based_datasource import DefaultBlockWritePathProvider
-
-    from awswrangler.distributed.ray.datasources import (  # pylint: disable=ungrouped-imports
-        PandasCSVDataSource,
-        PandasJSONDatasource,
-        PandasTextDatasource,
-        UserProvidedKeyBlockWritePathProvider,
-    )
-
-    if config.memory_format == MemoryFormat.MODIN.value:
-        from modin.pandas import DataFrame as ModinDataFrame
 
 _logger: logging.Logger = logging.getLogger(__name__)
 
@@ -84,81 +71,6 @@ def _to_text(  # pylint: disable=unused-argument
         elif file_format == "json":
             df.to_json(f, **pandas_kwargs)
     return [file_path]
-
-
-def _to_text_distributed(  # pylint: disable=unused-argument
-    df: pd.DataFrame,
-    file_format: str,
-    use_threads: Union[bool, int],
-    boto3_session: Optional[boto3.Session],
-    s3_additional_kwargs: Optional[Dict[str, str]],
-    path: Optional[str] = None,
-    path_root: Optional[str] = None,
-    filename_prefix: Optional[str] = uuid.uuid4().hex,
-    bucketing: bool = False,
-    **pandas_kwargs: Any,
-) -> List[str]:
-    if df.empty is True:
-        raise exceptions.EmptyDataFrame("DataFrame cannot be empty.")
-
-    if bucketing:
-        # Add bucket id to the prefix
-        prefix = f"{filename_prefix}_bucket-{df.name:05d}"
-        extension = f"{file_format}{_COMPRESSION_2_EXT.get(pandas_kwargs.get('compression'))}"
-
-        file_path = f"{path_root}{prefix}.{extension}"
-        path = file_path
-
-    elif path is None and path_root is not None:
-        file_path = path_root
-    elif path is not None and path_root is None:
-        file_path = path
-    else:
-        raise RuntimeError("path and path_root received at the same time.")
-
-    # Create Ray Dataset
-    ray_dataset = from_modin(df) if isinstance(df, ModinDataFrame) else from_pandas(df)
-
-    # Repartition into a single block if or writing into a single key or if bucketing is enabled
-    if ray_dataset.count() > 0 and path:
-        ray_dataset = ray_dataset.repartition(1)
-        _logger.warning(
-            "Repartitioning frame to single partition as a strict path was defined: %s. "
-            "This operation is inefficient for large datasets.",
-            path,
-        )
-
-    def _datasource_for_format(file_format: str) -> PandasTextDatasource:
-        if file_format == "csv":
-            return PandasCSVDataSource()
-
-        if file_format == "json":
-            return PandasJSONDatasource()
-
-        raise RuntimeError(f"Unknown file format: {file_format}")
-
-    datasource = _datasource_for_format(file_format)
-
-    mode, encoding, newline = _get_write_details(path=file_path, pandas_kwargs=pandas_kwargs)
-    ray_dataset.write_datasource(
-        datasource=datasource,
-        path=file_path,
-        block_path_provider=(
-            UserProvidedKeyBlockWritePathProvider()
-            if path and not path.endswith("/")
-            else DefaultBlockWritePathProvider()
-        ),
-        file_path=file_path,
-        dataset_uuid=filename_prefix,
-        boto3_session=None,
-        s3_additional_kwargs=s3_additional_kwargs,
-        mode=mode,
-        encoding=encoding,
-        newline=newline,
-        pandas_kwargs=pandas_kwargs,
-    )
-
-    return datasource.get_write_paths()
 
 
 @apply_configs
@@ -525,7 +437,7 @@ def to_csv(  # pylint: disable=too-many-arguments,too-many-locals,too-many-state
         description=description,
         parameters=parameters,
         columns_comments=columns_comments,
-        execution_engine=config.execution_engine,
+        execution_engine=engine.get(),
     )
 
     # Initializing defaults
@@ -959,7 +871,7 @@ def to_json(  # pylint: disable=too-many-arguments,too-many-locals,too-many-stat
         description=description,
         parameters=parameters,
         columns_comments=columns_comments,
-        execution_engine=config.execution_engine,
+        execution_engine=engine.get(),
     )
 
     # Initializing defaults
