+++ tests/unit/test_s3_select.py
@@ -3,9 +3,9 @@ import logging
 import pytest
 
 import awswrangler as wr
-from awswrangler._config import MemoryFormat
+from awswrangler._distributed import EngineEnum, MemoryFormatEnum
 
-if wr.config.memory_format == MemoryFormat.MODIN.value:
+if wr.engine.get() == EngineEnum.RAY.value and wr.memory_format.get() == MemoryFormatEnum.MODIN.value:
     import modin.pandas as pd
 else:
     import pandas as pd
